backend/requirements.txt
concurrent_log_handler==0.9.25
fastapi==0.110.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
psutil==5.9.8
SQLAlchemy==2.0.29
SQLAlchemy_serializer==1.4.12
starlette==0.37.2
uvicorn==0.29.0
aiosqlite==0.20.0
gunicorn==23.0.0
Cython==3.0.10
tensorflow==2.17.0
pillow==10.3.0